+++ docs/analysis/EigenvalueTrajectory.rst
@@ -13,13 +13,13 @@ first state in the trajectory is the ith eigen pair generated by the first diago
 
 State Tracking
 ---------------
-At each diagonalization of the CAP Hamiltonian, a set of n :math:`\eta`-dependent eigenpairs
-is generated, and the next point along the eigenvalue trajectory is chosen from that set 
-using one of two criterion: overlap and energy.
+At each diagonalization of the CAP Hamiltonian, left and right :math:`\eta`-dependent eigenpairs
+are computed and bi-orthogonalized. States are tracked using using one of two criterion: *overlap* and *energy*, which 
+is controlled by the **tracking** keyword argument of :func:`~pyopencap.analysis.CAPHamiltonian.track_state`.
 
 When overlap tracking is used (the default), at each step, the state with maximum overlap with the 
 previous state is chosen as the next point on the trajectory. The overlap is simply 
-calculated as the absolute value of the dot product between the eigenvectors. 
+calculated as the absolute value of the dot product between the right eigenvectors. 
 
 When energy tracking is used, at each step, the state with the minimum difference in energy 
 from the previous state is chosen. Since the energies are complex, the difference is computed as 
@@ -37,15 +37,37 @@ Raw uncorrected energies obtained from diagonalizaiton of the CAP Hamiltonian ca
 CAP onset and basis set quality. Practitioners of CAP theory often report so called `corrected` 
 energies, the exact form of which may vary from publication to publication. 
 
-We implement two forms of 1st order corrections. The default is the 
-The form of correction 
-we utilize here is that of [Cederbaum2002]_, which has the form:
+We implement two forms of 1st-order corrections: *density* and *derivative*, which is controlled by the **correction** 
+keyword argument of :func:`~pyopencap.analysis.CAPHamiltonian.track_state`. 
+
+The default is the density matrix correction of [Jagau2014]_ :
+
+:math:`U(\eta)=E(\eta)+ i \eta Tr [\gamma(\eta) W]`,
+
+where the trace expression is evaluated using the matrix elements of the CAP in the correlated basis:
+
+:math:`Tr [\gamma(\eta) W] = \sum_{kl} c^L_k(\eta)c^R_l(\eta)W_{kl}^{CB}`,
+
+and :math:`c^L` and :math:`c^R` xrefer to the components of the bi-orthogonalized left and right eigenvectors respectively.
+
+One can also use the first-order correction of [Cederbaum2002]_, which has the form:
 
 :math:`U(\eta)=E(\eta)-\eta\frac{\partial E(\eta) }{\partial \eta}`.
 
+The two approaches can be related to one another by the Hellman-Feynman theoreom, and in our experience yield 
+nearly identical results.
+
 Corrected trajectories are automatically computed when one obtains an :class:`~pyopencap.analysis.EigenvalueTrajectory`
-object generated by the :func:`~pyopencap.analysis.CAPHamiltonian.track_state` funciton, and 
+object generated by the :func:`~pyopencap.analysis.CAPHamiltonian.track_state` function, and 
 are stored in the class attribute `corrected_energies`.
+
+.. code-block:: python
+
+	# the various options for state tracking and corrections
+	traj = CAPH.track_state(1,tracking="energy",correction="density")
+	traj = CAPH.track_state(1,tracking="overlap",correction="derivative")
+
+
 
 :math:`\eta_{opt}`
 ------------------
@@ -60,10 +82,7 @@ For corrected trajectories, :math:`\eta_{opt}` is calculated as
 
 :math:`min |\eta\frac{dU}{d\eta}|`
 
-The derivative is calculated numerically by means of finite differences using the gradient 
-function of numpy_.
-
-.. _numpy: https://numpy.org/doc/stable/reference/generated/numpy.gradient.html
+The derivative is calculated numerically by means of finite differences.
 
 .. code-block:: python
 
@@ -112,6 +131,7 @@ for each point along the trajectory.
 References
 -----------
 .. [Cederbaum2002] Santra, R.; Cederbaum, L. S. Non-Hermitian Electronic Theory and Applications to Clusters. *Phys. Rep.* **2002**, 368 (1), 1–117.
+.. [Jagau2014] Jagau TC, Zuev D, Bravaya KB, Epifanovsky E, Krylov AI. A Fresh Look at Resonances and Complex Absorbing Potentials: Density Matrix-Based Approach. *J Phys Chem Lett.* **2014** 5, 2, 310–315
 
 
 .. autoclass:: pyopencap.analysis.EigenvalueTrajectory
