+++ pyopencap/analysis/CAPTrajectory.py
@@ -60,8 +60,10 @@ class Root():
         CAP strength parameter
     energy: float
         Total energy of state
-    eigvc: list of float
-        Eigenvector for state
+    Reigvc: list of float
+        Bi-orthogonalized right eigenvector for state
+    Leigvc: list of float
+        Bi-orthogonalized left eigenvector for state
     '''
     def __init__(self, energy, eta, Reigvc, Leigvc):
         '''
@@ -78,7 +80,7 @@ class Root():
         '''
         self.eta = eta
         self.energy = energy
-        self.eigvc = Reigvc
+        self.Reigvc = Reigvc
         self.Leigvc = Leigvc
 
 
@@ -87,8 +89,8 @@ class EigenvalueTrajectory():
     '''
     Eigenvalue trajectory generated by repeated diagonalizations of CAP Hamiltonian over range of eta values. 
 
-    States are tracked using either eigenvector overlap or energy criterion. Corrected energies(U) are obtained as:
-    U = E - eta*dE/deta.
+    States are tracked using either eigenvector overlap or energy criterion. Corrected energies are obtained using density matrix 
+    or first order correction.
     
     Attributes
     ----------
@@ -98,16 +100,16 @@ class EigenvalueTrajectory():
         List of uncorrected energies in trajectory.
     corrected_energies: list of float
         List of corrected energies in trajectory.
-    H0: np.ndarray of float: default=None
-        Zeroth order Hamiltonian in state basis
     W: np.ndarray of float: default=None
         CAP matrix in state basis. -1 prefactor is assumed.
     etas: list of float
         List of CAP strengths in trajectory.
-    correction: str
-        Choice of correction scheme. Either derivative (ref) or density (ref)
+    tracking: str, default="overlap"
+        Method to use to track the state
+    correction: str, default="density"
+        Choice of correction scheme. Either "density" or "derivative".
     '''
-    def __init__(self, state_idx, init_roots, H0, W, tracking="overlap", correction="density", pc=None):
+    def __init__(self, state_idx, init_roots, W, tracking="overlap", correction="density"):
         '''
         Initializes EigenvalueTrajectory object, which tracks a state starting from the first diagonalization at eta = 0.
         
@@ -117,14 +119,17 @@ class EigenvalueTrajectory():
             Index of state to track
         init_roots: list of :class:`~pyopencap.analysis.Root`
             Initial set of roots generated by diagonalization at eta = 0
+        W: np.ndarray of float: default=None
+            CAP matrix in state basis. -1 prefactor is assumed.
         tracking: str, default="overlap"
             Method to use to track the state
+        correction: str, default="density"
+            Choice of correction scheme. Either "density" or "derivative".
         '''
         self._last = init_roots[state_idx]
         self.states = [self._last]
         self.uncorrected_energies = [self._last.energy]
         self.corrected_energies = []
-        self.H0 = H0
         self.W = W
         self.etas = [0.0]
         if tracking == "overlap":
@@ -146,7 +151,7 @@ class EigenvalueTrajectory():
         maxo = 0.0
         state_idx = -1
         for st in states:
-            ov = np.abs(np.dot(st.eigvc, self._last.eigvc))
+            ov = np.abs(np.dot(st.Reigvc, self._last.Reigvc))
             if ov > maxo:
                 maxo = ov
                 cur = st
@@ -169,7 +174,7 @@ class EigenvalueTrajectory():
 
     def _density_matrix_correction(self):
         for i in range(0, len(self.states)):
-            eigvc = self.states[i].eigvc
+            eigvc = self.states[i].Reigvc
             Leigvc = self.states[i].Leigvc
             total = 0
             for k in range(0,len(self.W)):
@@ -697,6 +702,8 @@ class CAPHamiltonian():
         tracking: str, default="overlap"
             Method to use to track the state. Options are "overlap", which tracks based on eigenvector overlap, and "energy" which 
             tracks based on energy.
+        correction: str, default="density"
+            Choice of correction scheme. Either "density" or "derivative".
 
         Returns
         -------
@@ -707,7 +714,7 @@ class CAPHamiltonian():
             raise RuntimeError(
                 "Nothing to track. Execute `run_trajectory` first.")
         traj = EigenvalueTrajectory(state_idx,
-                                    self._all_roots[0],self.H0,self.W,tracking=tracking,correction=correction)
+                                    self._all_roots[0],self.W,tracking=tracking,correction=correction)
         for i in range(1, len(self._all_roots)):
             traj._add_state(self._all_roots[i])
         traj._calculate_corrected_energies()
