+++ pyopencap/tests/pyscf_test.py
@@ -19,33 +19,39 @@
     SOFTWARE.'''
 
 import pyopencap
-from pyscf import gto, scf, fci, tools
 import numpy as np
 import os
 import sys
+import pytest
 
-sys_dict = {
-    "molecule":"molden",
-    "basis_file":"molden_file.molden"
-}
+@pytest.fixture(autouse=True)
+def run_before_and_after_tests():
+    from pyscf import gto, scf, fci, tools
+    sys_dict = {
+        "molecule":"molden",
+        "basis_file":"molden_file.molden"
+    }
 
-cap_dict = {
-    "cap_type": "box",
-        "cap_x":"6.00",
-            "cap_y":"6.00",
-            "cap_z":"6.7",
-            "Radial_precision": "14",
-            "angular_points": "110"
-}
-mol = gto.M(
-            atom = 'H        0.0     0.0     0.54981512; \
-            H        0.0     0.0     -0.54981512;',
-            basis = "cc-pvdz"
-            )
-mol.build()
-myhf = scf.RHF(mol)
-myhf.kernel()
+    cap_dict = {
+        "cap_type": "box",
+            "cap_x":"6.00",
+                "cap_y":"6.00",
+                "cap_z":"6.7",
+                "Radial_precision": "14",
+                "angular_points": "110"
+    }
+    mol = gto.M(
+                atom = 'H        0.0     0.0     0.54981512; \
+                H        0.0     0.0     -0.54981512;',
+                basis = "cc-pvdz"
+                )
+    mol.build()
+    myhf = scf.RHF(mol)
+    myhf.kernel()
+    yield # this is where the testing happens
+    # Teardown : fill with any logic you want
 
+@pytest.mark.skipif('pyscf' not in sys.modules,reason="requires the pyscf library")
 def test_from_molden():
     tools.molden.from_scf(myhf,"molden_file.molden")
     s = pyopencap.System(sys_dict)
@@ -53,6 +59,7 @@ def test_from_molden():
     s.check_overlap_mat(pyscf_smat,"pyscf")
     os.remove("molden_file.molden")
 
+@pytest.mark.skipif('pyscf' not in sys.modules,reason="requires the pyscf library")
 def test_pyscf():
     tools.molden.from_scf(myhf,"molden_file.molden")
     s = pyopencap.System(sys_dict)
